+++ bin/common.mli
@@ -53,6 +53,8 @@ val term_with_default_root_is_cwd : t Cmdliner.Term.t
 (** Set whether Dune should print the "Entering directory '<dir>'" message *)
 val set_print_directory : t -> bool -> t
 
+val set_promote : t -> Dune_engine.Clflags.Promote.t -> t
+
 val debug_backtraces : bool Cmdliner.Term.t
 
 val config_from_config_file : Dune_config.Partial.t Cmdliner.Term.t